Standardize parameter order of fns that take the database as an argument
Rather than have a mix of fns that expect the DB as the first argument
and ones that expect it as the last argument in a map under the
`:connection` key we should standardize on the later.

Why? Because that's the format set by yesql which is by far the most
common type of database interacting fn we have. We shouldn't be
doing something different for the rare custom fn that needs the DB
as an argument, it's confusing.