Massive re-structuring to faciliate full assembly CCR/"best end date" calculations
The actual routing CCR calculation remains unchanged and, with mainly just
namespace refactoring, is being used "as-is" for the assembly calculation.

All of the "new" code is primarily concerned with retrieving the structure
details and writing the calculation results back to the database and then
wiring the new calculation into the core processing function alongside
the routes.