Set JVM timezone to UTC to avoid any 'helpful' date/time conversions
This was born out of the puzzling case of receiving 2015-04-08T23:00:00
for an end date instead of 2015-04-09(T00:00:00). Which spawned an
long investigative romp through time zones (see https://trello.com/c/Rl6dx3D4)
and has resulted in the creation of a PL/SQL package (lpe_timestamp_api)
and a set of guidelines to follow to avoid the issue in the future.