Moved route-ccrs.best-end-dates to .cljc
For some reason the multifn wasn't actually being populated with
all the methods in ClojureScript (regardless of using the defmethods
macro or just defmethod calls.) Something seemed to be causing
previous method definitions to get overwritten when adding new ones.

So, the function has been re-written to use `cond` for dispatch
instead. There's no real drawback to this as there are relatively
few dispatch values and it didn't need to be openly extensible.