Moved `sourced?` to schema namespace and made `route-ccrs.start-dates` a `.cljc`
Moving `sourced?` was the simplest way to separate it from the `defmethods`
macro in `route-ccrs.util`. Moving it also makes sense given that it looks
like that macro is not long destined for this world and it's really the
only reason that the `route-ccrs.util` namespace exists.

`start-date` has undergone a similar multi-method -> cond change as
`best-end-date`, for the same reasons.