Use types with value equality for generated struct/route IDs in tests
Running the schema tests would frequently fail due to the use of
"simple" types for the struct/route IDs which had the possibility
of generating values that did not satisfy an equality check. As a
consequence there was a lack of consistency in the test runs that
made it hard to tell whether the failures were transient errors
due to the generated values or actual test failures.